refactor: types into new package
