+++ src/main/java/li/zhongfu/cs2103/chatbot/types/ToDo.java
@@ -1,4 +1,4 @@
-package li.zhongfu.cs2103.chatbot;
+package li.zhongfu.cs2103.chatbot.types;
 
 public class ToDo extends Task {
     public ToDo(String name) {
